backend/pyproject.toml
[project]
name = "coasting"
version = "0.1.0"
description = "A project for splitting an image to print on stackable coasters"
authors = [
    {name = "Nick Sullivan", email = "nick.dave.sullivan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.4",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0",
    "pillow>=12.1.0",
    "python-multipart>=0.0.20",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.1.0",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coasting"]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"

[tool.uv]
dev-dependencies = ["ruff>=0.11.0"]